Have added authentication, login, logout with functions from django.contrib.auth; have also added functioning links on the tiles and wall page, where there were previously only placeholders

+++ tiles/views.py
@@ -1,24 +1,25 @@
 from django.shortcuts import render
 from django.http import HttpResponse
 from django.shortcuts import redirect
-from .forms import TileForm
-from .models import User, Tile
+from django.utils import timezone
+from .forms import TileForm, RegForm
+from .models import Tile
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+
 
 user = {'id': 2, 'username': 'Anush', 'score':-1}
 
 tiles = [{'id': 1, 'created': 'today', 'content': 'I am very sad'},
 		{'id': 2, 'created': 'today',  'content': 'Why do I not manage to do things properly?'},
 		{'id': 3, 'created': 'today', 'content': 'Why do I not have anything to show?'},
-		{'id': 4, 'created': 'today',  'content': 'I really want to accomplish things but seem to lack motivation'},
-]
+		{'id': 4, 'created': 'today',  'content': 'I really want to accomplish things but seem to lack motivation'},]
 
 # Create your views here.
-def welcome(request):
-	return render(request,'welcome.html',{})
+
 
 def test(request):
-	return render(request,'test.html',{})
-
+	return render(request,'tiles/test.html',{})
 
 def myTiles(request,user_id):
 	if request.method == "POST":
@@ -33,21 +34,38 @@ def myTiles(request,user_id):
 		user = User.objects.filter(id = user_id)
 		tiles = Tile.objects.filter(author_id=user_id)
 		if user:
-			return render(request,'tiles.html',{'tiles':tiles, 
+			return render(request,'tiles/tiles.html',{'tiles':tiles, 
 												'user': user[0],
 												 'form': form})
 		else:
 			return HttpResponse('This user does not exist!')
 
 def deleteTile(request,user_id,tile_id):
-	user = User.objects.get(id = user_id)
-	tile = Tile.objects.get(id=tile_id)
-	tile.delete()
-	tiles = Tile.objects.all()
-	return redirect('myTiles',user_id=user.id)
+
+	if request.method == "POST":
+		user = User.objects.get(id = user_id)
+		tile = Tile.objects.get(id=tile_id)
+		tile.delete()
+		return redirect('myTiles',user_id=user.id)
+
+	else:
+		return render(request,'tiles/tile_delete.html',{})
 
 def editTile(request,user_id,tile_id):
-	return HttpResponse('Edit page')
+	tile = Tile.objects.get(id=tile_id)
+	
+	if request.method == "POST":
+		form = TileForm(request.POST,instance = tile)
+		if form.is_valid():
+			tile = form.save(commit=False)
+			tile.author = User.objects.get(id = user_id)
+			tile.created_date = timezone.now()
+			tile.save()
+			return redirect('myTiles', user_id=user_id)
+		return redirect('editTile',{'form':form})
+	else:
+		form = TileForm(instance = tile)
+		return render(request,'tiles/tile_edit.html',{'form':form})
 
 def shareTile(request,user_id,tile_id):
 	user = User.objects.get(id = user_id)
@@ -70,8 +88,6 @@ def tileSolved(request,user_id,tile_id):
 	tile.save()
 	return redirect('myTiles',user_id=user.id)
 
-
-
 from django.http import JsonResponse
 import random
 def ajaxTest(request):
@@ -80,3 +96,80 @@ def ajaxTest(request):
 
 def settings(request,user_id):
 	return HttpResponse('Your settings')
+
+
+
+"""def login(request):
+	form = RegForm()
+	if request.method == "POST":
+		
+		username = request.POST['username']
+		password = request.POST['password']
+		print (username)
+		#username = form.username
+		#password = form.password
+		user = authenticate(username=username, password=password)
+		print (user)
+		if user is not None:
+			
+			if user.is_active:
+				login(request, user)
+				return redirect('myTiles',user_id=user.id)
+		return render(request,'tiles/welcome.html',{'form': form})
+
+	else:
+		form = RegForm()
+		return render(request,'tiles/welcome.html',{'form': form})"""
+
+def welcome(request):
+	if request.method == "POST":
+		form = RegForm(request.POST)
+		if form.is_valid():
+			username = request.POST['username']
+			password = request.POST['password']
+			u = User.objects.create_user(username = username, password = password)
+
+			return redirect('myTiles',user_id=u.id)
+		"""	
+		user = authenticate(username=username, password=password)
+		print (user)
+		if user is not None:
+			
+			if user.is_active:
+				login(request, user)
+				return redirect('myTiles',user_id=user.id)
+		"""
+		return render(request,'tiles/welcome.html',{'form': form})
+
+	else:
+		form = RegForm()
+		return render(request,'tiles/welcome.html',{'form': form})
+
+def wallLogin(request):
+	if request.method == "POST":
+		form = RegForm()
+		##form = RegForm(request.POST)
+		##if form.is_valid():
+		print ('authenticate called\n')
+		username = request.POST['uname']
+		password = request.POST['pword']
+		
+		user = authenticate(username=username, password=password)
+		
+		if user is not None:
+			print('Successfully authenticated\n')
+			if user.is_active:
+				
+				login(request, user)
+				print('Successfully logged in\n')
+				return redirect('myTiles',user_id=user.id)
+		#return render(request,'tiles/login.html')
+		return render(request,'tiles/login.html',{'error' :'You have not correctly entered your username and/or password'})
+	else:
+		return render(request,'tiles/login.html', {'error': ''})
+
+
+def wallLogout(request):
+	logout(request)
+	return redirect('welcome')
+
